Add list screen

+++ app/src/main/java/com/example/jetpackcomposeproject/data/dao/AthleteDAO.kt
@@ -18,7 +18,7 @@ interface AthleteDAO {
     fun insert(item: Athlete): Long
 
     @Update
-    fun update(item: Athlete): Long
+    fun update(item: Athlete): Int
 
     @Delete
     fun delete(item: Athlete)
